Grunt, Notes css, and file updates
updated the grunt file so it wouldn't run criticalcss on the watch task anymore. Also modified the Notes CSS a bit to give larger margins to the left and right for better readability (will have to test with real fonts. Added HTML files for the other proposed notes as well as changed a bit of markup to be more semantic